onu/requirements.txt
asgiref==3.6.0
Django==3.2.13
djangorestframework==3.14.0
mysqlclient==2.1.1
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4

pandas~=2.0.1
numpy~=1.24.3
scikit-learn~=1.2.2

